Sixth iOS App
This app is based of the app Mark instructs on Udemy. I just changed
the artist. (Daughtry is one of my favorite musical artist.) I learned
how to bring videos into an app and how to make them play. The thing
is, you need to click on the title of the video for it to play.